Give AYTQ and Check separate error controllers
Rather than sharing a single custom error controller, give each service
its own. This will allow us to use the correct layout templates and
helps ensure the content on a given error page is specific to the
service it is rendered from.
