+++ app/views/check_records/errors/internal_server_error.html.erb
