+++ app/views/check_records/errors/not_found.html.erb
