+++ app/views/qualifications/errors/internal_server_error.html.erb
@@ -1,0 +1,18 @@
+<%= content_for :page_title, "Sorry, there’s a problem with the service" %>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+    <h1 class="govuk-heading-l">Sorry, there’s a problem with the service</h1>
+
+    <p class="govuk-body">Try again later.</p>
+
+    <p class="govuk-body">
+      If you reached this page after submitting information then it has not
+      been saved. You’ll need to enter it again when the service is available.
+    </p>
+    <p class="govuk-body">
+      If you have any questions, please email us at <a class="govuk-link"
+        href="mailto:<%= t('service.email') %>"><%= t('service.email') %></a>.
+    </p>
+  </div>
+</div>
