+++ app/views/qualifications/errors/not_found.html.erb
@@ -1,0 +1,19 @@
+<%= content_for :page_title, "Page not found" %>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+    <h1 class="govuk-heading-l">Page not found</h1>
+    <p class="govuk-body">
+      If you entered a web address, check it is correct.
+    </p>
+    <p class="govuk-body">
+      If you pasted the web address, check you copied the entire address.
+    </p>
+    <p class="govuk-body">
+      If the web address is correct or you selected a link or button and you
+      need to speak to someone about this problem, contact the <%=
+      t('service.name') %> team: <a class="govuk-link" href="mailto:<%=
+      t('service.email') %>"><%= t('service.email') %></a>.
+    </p>
+  </div>
+</div>
