+++ app/views/qualifications/errors/unprocessable_entity.html.erb
@@ -1,0 +1,15 @@
+<%= content_for :page_title, 'Sorry, there’s a problem with the service' %>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+    <h1 class="govuk-heading-l">Sorry, there’s a problem with the service</h1>
+
+    <p class="govuk-body">Try again later.</p>
+
+    <p class="govuk-body">
+      If you continue to see this error contact the <%= t('service.name')
+      %> team: <a class="govuk-link" href="mailto:<%= t('service.email') %>"><%=
+      t('service.email') %></a>.
+    </p>
+  </div>
+</div>
