+++ config/routes.rb
@@ -32,10 +32,4 @@ Rails.application.routes.draw do
 
   constraints(RouteConstraints::AccessYourTeachingQualificationsConstraint.new) { draw(:aytq) }
   constraints(RouteConstraints::CheckRecordsConstraint.new) { draw(:check_records) }
-
-  scope via: :all do
-    get '/404', to: 'errors#not_found'
-    get '/422', to: 'errors#unprocessable_entity'
-    get '/500', to: 'errors#internal_server_error'
-  end
 end
