+++ config/routes/aytq.rb
@@ -21,4 +21,10 @@ namespace :qualifications do
   end
 end
 
+scope via: :all do
+  get '/404', to: 'qualifications/errors#not_found'
+  get '/422', to: 'qualifications/errors#unprocessable_entity'
+  get '/500', to: 'qualifications/errors#internal_server_error'
+end
+
 root to: redirect("/qualifications/start"), as: :qualifications_root
